Fix reference issues
