+++ src/Factories/Issue/index.ts
@@ -1,7 +1,8 @@
 import { IFactory } from "../index";
 import { v4 as uuid } from 'uuid';
 import { IssueType } from '../../Domain/IssueType';
-import { IssueLocation, BlockLocation } from '../../Domain/IssueLocation';
+import { BlockLocation } from "../../Domain/BlockLocation";
+import { IssueLocation } from "../../Domain/IssueLocation";
 import { IIssue, Issue } from '../../Domain/Issues';
 
 export class IssueFactory implements IFactory<IIssue>  {
